Changed from static to instance
This was necessary to be able to add LastLoadWasSuccessful. This property can be queried to test if a mod file could be loaded. The correct behaviour in this case is usually not applying any mods, because the returned mod-configuration will use default values that are usually not the expected values in game.